third_party/blink/web_tests/platform/generic/fast/forms/select/listbox-selection-2-expected.txt

<select> selection test for scrolling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selectionPattern("sl1") is "10000000000000000"
PASS selectionPattern("sl1") is "01000000000000000"
PASS selectionPattern("sl1") is "00000010000000000"
PASS selectionPattern("sl1") is "00000001000000000"
PASS selectionPattern("sl1") is "10000000000000000"
PASS selectionPattern("sl1") is "00000001000000000"
PASS selectionPattern("sl2") is "010"
PASS selectionPattern("sl2") is "010"
PASS selectionPattern("sl2") is "001"
PASS successfullyParsed is true

TEST COMPLETE

one
two
three
four
five
six
seven
eight
nine
ten
eleven
twelve
thirteen
fourteen
fifteen
sixteen
seventeen
one
two
three
